Move restructureForCompliance functions to correct client script
restructureForCompliance.py contained several functions that were not
used in that script, but were imported into other client scripts.  Move
those functions to the correct client script.  Add common code to
archivematicaFunctions and import that.
